// BB-main/src/bb/generated/BbPhaseSweepResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * 위상 스윕 결과 (D-8). `BbPhaseSweep::enabled` 일 때만 채워진다.
 */
export type BbPhaseSweepResult = { 
/**
 * 최악 Q_max [N] 와 그때의 φ₀ [rad]
 */
worst_q_max_n: number, worst_q_max_phase_rad: number, 
/**
 * 최악 최대접촉응력 [MPa] 와 그때의 φ₀ [rad]
 */
worst_p_max_mpa: number, worst_p_max_phase_rad: number, 
/**
 * (φ₀, Q_max) 전 이력
 */
curve: Array<[number, number]>, };